ciao-sdk: Add new package to contain ciao library
Ciao should have a library that the command line can call into to do
various actions within the cluster. This is separated out of the CLI,
so that any package or other code can call it to get information
or send requests to the controller or compute endpoints. This also
enables other entities to interact with the cluster without having
to call the CLI at all.

Signed-off-by: Tudor Marcu <tudor.marcu@intel.com>