ciao-sdk: Split up individual objects and API
To more logically separate things, each object i.e workload, instance,
etc, will have its own file to contain the functionality it provides and
supports. The CLI code is fully contained within the ciao/ folder, with
the ciao-sdk/ folder containing stricly API implementation, with no
knoweldge of the cobra package or types. The intent is to provide a more
generic implementation, to do things such as:
- Show("workload", nil) // Display all workloads
Which does not require the CLI to pass in options.

Signed-off-by: Tudor Marcu <tudor.marcu@intel.com>