ciao-controller: Log whenever a request generates an error
When a request generates an error log this via glog. Now that we are
better at handling errors thoughout controller this change means that
all errors generated from requests will be logged into the controller
log. This should provide more details than the error reported to the
user as that is the result of conversion to a HTTP error code.

Signed-off-by: Rob Bradford <robert.bradford@intel.com>