Update library to be more user friendly
We shouldn't expect users to know every detail of the ciao data structs
and how to handle them correctly, even with the client library available
for many of the functions. This attempts to abstract it a little more
and provide library functions that return more usable data to be used in
various different contexts.

Signed-off-by: Tudor Marcu <tudor.marcu@intel.com>