Add node list, and pretty-print for output
The tfortools template printing is used for all listing/show functions,
and to make the output more usable and applicable to the library, it
will write to a bytes.Buffer instead of stdout. This can then be used
and processed as seen fit without forcing a print from any functions.

Signed-off-by: Tudor Marcu <tudor.marcu@intel.com>